% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trytrytry.R
\name{tryTryTry}
\alias{tryTryTry}
\title{Try an expression n times before quitting}
\usage{
tryTryTry(x, n = 3L)
}
\arguments{
\item{x}{The expression to try.}

\item{n}{The number of times to try. Default is 3.}
}
\description{
Try an expression n times before quitting
}
